+++ src/api/dart/api/room.rs
@@ -4,14 +4,14 @@ use flutter_rust_bridge::{DartOpaque, frb};
 use send_wrapper::SendWrapper;
 use tracerr::Traced;
 
+#[cfg(doc)]
+use crate::{api::LocalMediaTrack, room::Room};
 use crate::{
     api::{Error as DartError, api::ApiMediaStreamSettings},
     media::MediaSourceKind,
     platform::{self, utils::dart_future::IntoDartFuture as _},
     room as core,
 };
-#[cfg(doc)]
-use crate::{media::track::local::LocalMediaTrack, room::Room};
 
 /// External handle to a [`Room`].
 #[derive(Debug)]
@@ -361,8 +361,8 @@ impl RoomHandle {
             .map_err(Into::into)
     }
 
-    /// Sets a callback to be invoked once a new [`LocalMediaTrack`] is added
-    /// to the provided [`Room`].
+    /// Sets a callback to be invoked once a new [`LocalMediaTrack`] is
+    /// added to the provided [`Room`].
     ///
     /// This might happen in such cases:
     /// 1. Media server initiates a media request.